% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_panel.R
\name{create_panel}
\alias{create_panel}
\title{Create a panel of dates}
\usage{
create_panel(start_date, end_date, by)
}
\arguments{
\item{start_date}{A character in YYYY-MM-DD format. The start date of the panel}

\item{end_date}{A character in YYYY-MM-DD format. The end date of the panel}

\item{by}{}
}
\value{
A tibble of dates in panel-format
}
\description{
`create_panel` creates a panel of dates.
}
\examples{

##
create_panel()



}
