% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_false_positive_dates.R
\name{filter_false_positive_dates}
\alias{filter_false_positive_dates}
\title{Pass a panelsummary::panelsummary_raw dataframe into kableExtra::kbl() with typical defaults}
\usage{
filter_false_positive_dates(x)
}
\arguments{
\item{x}{The data.frame (or tibble) from that needs the dates removed.}
}
\value{
A raw data frame that is ready for further manipulation.
}
\description{
`filter_false_positive_dates` omits any dates that are considered to be high false positives.
}
\examples{

##
filter_false_positive_dates(x)



}
